Node Weighting, StreamService will use FilterQuery instead of GeoFilter, Modded model

+++ app/src/main/java/com/dankideacentral/dic/model/WeightedNode.java
@@ -7,6 +7,6 @@ import com.google.maps.android.clustering.ClusterItem;
  */
 
 public abstract class WeightedNode implements ClusterItem {
-    public abstract float getWeight();
+    public abstract int getWeight();
 }
